update route use middleware auth and bug user product

+++ routes/web.php
@@ -46,19 +46,19 @@ Route::get('/home', 'IndexController@index')->name('home');
 
 Route::get('admin/customer', 'CustomerController@index')->name('appshell.customer.index');
 
-Route::get('customer/product', 'CustomerProductController@index')->name('customer.product.index');
-Route::get('customer/product/{product}', 'CustomerProductController@show')->name('customer.product.show');
-Route::get('customer/product/create', 'CustomerProductController@create')->name('customer.product.create');
-Route::post('customer/product/store', 'CustomerProductController@store')->name('customer.product.store');
-Route::get('customer/product/{product}/edit', 'CustomerProductController@edit')->name('customer.product.edit');
-Route::put('customer/product/{product}', 'CustomerProductController@update')->name('customer.product.update');
-Route::delete('customer/product/{product}', 'CustomerProductController@destroy')->name('customer.product.destroy');
+Route::get('customer/product', 'CustomerProductController@index')->middleware('auth')->name('customer.product.index');
+Route::get('customer/product/{product}', 'CustomerProductController@show')->middleware('auth')->name('customer.product.show');
+Route::get('customer/product/create', 'CustomerProductController@create')->middleware('auth')->name('customer.product.create');
+Route::post('customer/product/store', 'CustomerProductController@store')->middleware('auth')->name('customer.product.store');
+Route::get('customer/product/{product}/edit', 'CustomerProductController@edit')->middleware('auth')->name('customer.product.edit');
+Route::put('customer/product/{product}', 'CustomerProductController@update')->middleware('auth')->name('customer.product.update');
+Route::delete('customer/product/{product}', 'CustomerProductController@destroy')->middleware('auth')->name('customer.product.destroy');
 
 Route::get('profile/{id}', 'UserProfileController@show');
 
-Route::get('user_profile/{id}', 'UserProfileController@showUser');
-Route::post('user_profile/{id}', 'UserProfileController@update');
-Route::post('avatar_profile', 'UserProfileController@update_avatar');
+Route::get('user_profile/{id}', 'UserProfileController@showUser')->middleware('auth');
+Route::post('user_profile/{id}', 'UserProfileController@update')->middleware('auth');
+Route::post('avatar_profile', 'UserProfileController@update_avatar')->middleware('auth');
 
 
 // Route::resource('customer', 'CustomerController', ['names' => [
